api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pystray==0.19.5
Pillow==10.1.0


